[DCIP-739] Add tandem boot changes for dpdk driver

+++ drivers/dpdk/drivers/net/qdma/rte_pmd_qdma.h
@@ -587,7 +587,7 @@ int rte_pmd_qdma_set_c2h_descriptor_prefetch(int port_id, uint32_t qid,
  * @ingroup rte_pmd_qdma_func
  *****************************************************************************/
 int rte_pmd_qdma_set_mm_endpoint_addr(int port_id, uint32_t qid,
-			enum rte_pmd_qdma_dir_type dir, uint32_t addr);
+			enum rte_pmd_qdma_dir_type dir, uint64_t addr);
 
 /******************************************************************************/
 /**
